\name{nsga2NREL-package}
\alias{nsga2NREL-package}
\docType{package}
\title{
Elitist Non-dominated Sorting Genetic Algorithm based on R
}
\description{
Functions for box-constrained multiobjective optimization using the elitist non-dominated sorting genetic algorithm - NSGA-II.
}
\details{
\tabular{ll}{
Package: \tab nsga2NREL\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-12\cr
License: \tab LGPL-3\cr
}
This package provide functions for box-constrained multiobjective optimization using the elitist non-dominated sorting genetic algorithm - NSGA-II. Fast non-dominated sorting, crowding distance, tournament selection, simulated binary crossover, and polynomial mutation are called in the main program, nsga2R, to complete the search.
}
\author{
Ching-Shih (Vince) Tsou <cstsou@mail.ntcb.edu.tw>

Maintainer: Ming-Chang (Alan) Lee <alan9956@ydu.edu.tw>
}
\references{
Deb, K., Pratap, A., Agarwal, S., and Meyarivan, T. (2002), " A fast and elitist multiobjective genetic algorithm: NSGA-II", IEEE Transactions on Evolutionary Computation, 6(2), 182-197.
}

